+++ src/components/Cards/PersonCard/PersonCard.stories.js
@@ -3,7 +3,7 @@ import React from 'react';
 import PersonCard from './index';
 
 export default {
-  title: 'Components|Person/PersonCard',
+  title: 'Components|Cards/PersonCard',
   component: PersonCard,
 };
