+++ src/components/Cards/shared/PersonHeadShot/PersonHeadShot.stories.js
@@ -3,7 +3,7 @@ import React from 'react';
 import PersonHeadShot from './index';
 
 export default {
-  title: 'Components|Person/PersonHeadShot',
+  title: 'Components|Cards/shared/PersonHeadShot',
   component: PersonHeadShot,
 };
